executor: split test file.
